Add files via upload
Added a invariant so that the lengths in of all dimension of the arrays keys and vals are equal to each other. (About 15 minutes)
Added contracts for the resize method.  (About 30 minutes)
In the last commit i forgot to mention how long the contract updates took.  get : 30 minutes, put : 90 minutes, delete : 60 minutes
Added a short description to each contract.