Add files via upload
Added KeY statistics of the Hash Table concepts.  As odf spreadsheet and pdf.

Added a new put method to LiniearProbing, that stops when it finds a null in the table. (Specification is not complete)

Updated put and increaseArraySize, they can be verified in the current form. They are still incomplete.