Add files via upload
Updated the current Implementations (Normal and WithInt) and the WithInt works in the new simplified version.
Added two new versions of the two Concepts.
One uses int Keys, but Object Values.
The other HashObject Keys and Object Values, but uses == instead of the equals method from HashObject.
Also Updated the Statistics with all the Results that are possible. If the entrys are empty it means it is not able to finish. (Mostly because it take to long)